Add files via upload
Python script to convert the TraderVehicleParts.txt from the Dr. Jones Trader mod for day to a format compatible with the TraderPlusVehiclesConfig.json file for the TraderPlus mod. Currently creates 6 lines of repeated code at the end of the file but other than that it seems to work well enough to save some time.